feat:  configure consultation day mode on home page
