+++ dashboard/templates/index.html
@@ -228,12 +228,12 @@
                                 {% for time in 8|times:15 %}
                                     <tr class="{% cycle 'even-row' 'odd-row' %}" id="appointment-{{ forloop.counter }}" 
                                         {% for appointment in appointments %}
-                                            {% if appointment.appointment_date == next_consult_date.date and appointment.appointment_time|format_time == time %}
+                                            {% if appointment.date == next_consult_date.date and appointment.time|format_time == time %}
                                                 data-appointment-id="{{ appointment.id }}"
                                                 onclick="showPatientData('{{ appointment.id }}')"
                                             {% endif %}
                                         {% endfor %}>
-                                        <th>{{ time }}</th>
+                                        <th class="appointment-time">{{ time }}</th>
                                         <td>
                                             <button class="text-blue-500 flex items-center justify-center w-full h-full">
                                                 {% for appointment in appointments %}
@@ -349,9 +349,14 @@
     </div>
         <!-- end main content section -->
 </div>
+
 <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
 <script src="{% static "js/dashboard/index.js" %}"></script>
-<script src="{% static "js/dashboard/queue_appointments.js" %}"></script>
+<script src="{% static "js/dashboard/queueAppointments.js" %}"></script>
+<script src="{% static "js/dashboard/patientPanel.js" %}"></script>
+<script src="{% static "js/dashboard/setPaymentMehod.js" %}"></script>
+<script src="{% static "js/dashboard/checkHomeState.js" %}"></script>
+
 <script>
     function closeModal() {
         document.getElementById('addEventModal').classList.add('hidden');
