+++ dashboard/urls.py
@@ -4,9 +4,11 @@ from .  import views
 urlpatterns = [
     path('', views.home, name="home"),
     path('next-consult-date/save', views.save_new_consult_date, name='save_next_consult_date'),
+    path('start_queue/', views.start_queue, name='start_queue'),
+    path('update_last_treated_appointment/<int:appointment_id>/', views.update_last_treated_appointment, name='update_last_treated_appointment'),
+    path('check_queue_state/', views.check_queue_state, name='check_queue_status'),
 
     path('settings/', views.settings_view, name="settings"),
-    
     path('payment-methods/add', views.add_payment_method, name="add_payment_method"),
     path('payment-methods/delete/<int:id>/', views.delete_payment_method, name="delete_payment_method")
 ]
