+++ static/js/custom.js
@@ -7,7 +7,7 @@
         rtlClass: 'ltr', // rtl, ltr
         animation: '', // animate__fadeIn, animate__fadeInDown, animate__fadeInUp, animate__fadeInLeft, animate__fadeInRight, animate__slideInDown, animate__slideInLeft, animate__slideInRight, animate__zoomIn
         navbar: 'navbar-sticky', // navbar-sticky, navbar-floating, navbar-static
-        semidark: false,
+        semidark: true,
     };
     window.addEventListener('load', function () {
         // screen loader
